requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.6
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
Markdown==3.3.4
Pillow==8.3.1
pycodestyle==2.7.0
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
